backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
stripe
python-dotenv
alembic